refactor in packages to create flask app with multiple blueprints/sqlalchemy extensions correctly
