+++ pymityma/api/__init__.py
@@ -1,0 +1,2 @@
+from .app import create_app
+from .server import ma, migrate, db
