+++ pymityma/api/route/producer/routes.py
@@ -1,0 +1,6 @@
+from flask import Blueprint
+from .main import *
+
+producer_api = Blueprint('producer', __name__)
+producer_api.add_url_rule('/', view_func=producers, methods=['GET', 'POST'])
+producer_api.add_url_rule('/<id>', view_func=remove_producer, methods=['DELETE'])
