+++ pymityma/server.py
@@ -1,5 +1,5 @@
 import os
-from api.server import create_app
+from api import create_app
 
 app = create_app()
